Move icon components to /components/icons, rename services folder to fetches and add alt attribute for avatars.
