+++ components/features/auth/AuthPromoAside.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import PenwwwsIcon from "@/public/assets/PenwwwsIcon";
+import PenwwwsIcon from "@/components/icons/PenwwwsIcon";
 
 export default function AuthPromoAside() {
   return (
