+++ components/icons/PenwwwsIcon.tsx
