+++ fetches/school.ts
